Fix compile error splashscreen +  update CI to store html coverage

+++ src/splashScreen.py
@@ -222,7 +222,7 @@ if __name__ == "__main__":
 
     for file in zf.infolist():
         extracted_size += file.file_size
-        percentage = extracted_size/uncompress_size*100
+        percentage = int(extracted_size/uncompress_size*100)
         progressBar.setValue(percentage)
         if percentage > 50 and percentage < 70:
             splash.showMessage("Copying temp files", Qt.AlignBottom | Qt.AlignCenter, Qt.black)
